# [Bugfix] V1 Memory Profiling: V0 Sampler Integration without Rejection Sampler