--- a/vllm/v1/worker/gpu_model_runner.py
+++ b/vllm/v1/worker/gpu_model_runner.py
@@ -1304,8 +1304,8 @@ class GPUModelRunner(LoRAModelRunnerMixin):
             if get_pp_group().is_last_rank:
                 hidden_states = hidden_states[logit_indices]
                 logits = self.model.compute_logits(hidden_states, None)
-                penalties = lambda: torch.full(
-                    (num_reqs, ), 0.0, device=self.device)
+                dummy_tensors = lambda v: torch.full(
+                    (num_reqs, ), v, device=self.device)
                 dummy_metadata = SamplingMetadata(
                     temperature=torch.full((num_reqs, ),
                                            0.5,
@@ -1313,18 +1313,16 @@ class GPUModelRunner(LoRAModelRunnerMixin):
                     all_greedy=False,
                     all_random=False,
                     spec_token_ids=None,
-                    top_p=torch.full((num_reqs, ), 0.99, device=self.device),
-                    top_k=torch.full((num_reqs, ),
-                                     logits.size(1) - 1,
-                                     device=self.device),
+                    top_p=dummy_tensors(0.9),
+                    top_k=dummy_tensors(logits.size(1) - 1),
                     min_p=None,
                     generators={},
                     max_num_logprobs=None,
-                    no_penalties=True,
-                    prompt_token_ids=None,
-                    frequency_penalties=penalties(),
-                    presence_penalties=penalties(),
-                    repetition_penalties=penalties(),
+                    no_penalties=False,
+                    prompt_token_ids=torch.ones_like(logits, dtype=torch.long),
+                    frequency_penalties=dummy_tensors(0.1),
+                    presence_penalties=dummy_tensors(0.1),
+                    repetition_penalties=dummy_tensors(0.1),
                     output_token_ids=[[] for _ in range(num_reqs)],
                     min_tokens={},
                     logit_bias=[None for _ in range(num_reqs)])
@@ -1333,10 +1331,9 @@ class GPUModelRunner(LoRAModelRunnerMixin):
             else:
                 logits = None
                 sampler_output = None
-                penalties = None
                 dummy_metadata = None
             torch.cuda.synchronize()
-            del hidden_states, logits, sampler_output, penalties, dummy_metadata
+            del hidden_states, logits, sampler_output, dummy_metadata
             self.encoder_cache.clear()
         gc.collect()