@WoosukKwon 2025-02-22
@JenZhao Congrats for your first PR!

@JenZhao @ywang96 I've observed that the peak memory goes much beyond 90%, even after this PR:
```
[4] NVIDIA H100 80GB HBM3 | 64'C,  94 % | 80773 / 81559 MB | woosuk.kwon(80296M)
```
I encountered this error while running `benchmark_serving.py`. The only modification I made was changing the temperature parameter from 0 to 0.1.