@WoosukKwon 2025-02-22
BTW, weirdly, the memory taken by the vLLM instance got reduced when the server finished handling all the requests and became idle.
```
[4] NVIDIA H100 80GB HBM3 | 33'C,   0 % | 71009 / 81559 MB | woosuk.kwon(70532M)
```
As we do not call `torch.cuda.empty_cache` by ourselves, this is quite weird.